ui: integer UI scale; no floating point in interrupt handlers
The mouse IRQ handler clamped the cursor with vbe_get_width() / ui_scale
in x87 floating point (fdivs, fildl, fistpl and two control word
switches), in the middle of whatever task it interrupted, whose x87
state is not saved on interrupt entry.

ui_scale only ever is 1 or 2 now, so it becomes an int; every use gives
the same result as before (the one division by 2 in the font code stays
a float division). The files on the interrupt path - isr, scheduler,
event queue, PIT, PIC, keyboard, mouse, serial - are compiled with
-mgeneral-regs-only, so any new floating point use there is a compile
error. Their functions contain no x87 instructions (checked with
objdump; mouse_callback had 12 before).

+++ kernel/gfx/font/font.c
@@ -181,7 +181,7 @@ void font_draw_text(gfx_context_t *ctx, int x, int y, const char *text,
       pen_x += advance;
       tt_free_glyph(&glyph);
     } else {
-      pen_x += (size_px * ui_scale) / 2;
+      pen_x += (float)(size_px * ui_scale) / 2.0f;
     }
   }
 }
